python models: add basic SQLite models
These models should make it easier for us to iterate on cool statistics
and visualizations on the website. In addition, it should support any
future query types, searching, or filtering we wish to do.

These models are at best a first pass, and should not be taken as
final. However the method of constructing these models is non-
destructive, so it should be safe to iterate on them at any point.

Example use:
```
#!/usr/bin/env python3

import db

with db.create_or_open("baalor200.sqlite") as connection:
    db.create_tables_if_not_exists(connection)

    db.import_files_if_not_imported(connection, "Baalor200")

    print("---------------------------------------------------")
    cursor = connection.cursor()
    cursor.execute("SELECT key, COUNT(*) AS total FROM potionsObtained GROUP BY key ORDER BY total DESC;")
    print(cursor.getdescription())
    for x in cursor:
        print(x)
    print("---------------------------------------------------")
```