py runs: add archive link support to runs
This attempts to generate timestamp'd links to runs in the youtube
archive for runs. This does so by computing the amount of time
from the start of this run and the start of the first run for the day,
and using that as an additional offset into the youtube video.

This relies on the existance of a `data/youtube.json` file, which
must contain every archive videos id, title, and description. This
can be generated and maintained with a background worker configured
to run every hour.

Without the file, all runs should appear to have no link, and thus
the output of the website should be uneffected.