Docker compose for running the Python application -
Implement a bearbones Docker setup.

Simple configuration for the networking (`host`) and fs. The working directory for the Python app is mapped to the root of the repo, so in-theory there should be very minimal side-effects.

Also left some slots for future additions to the compose - the idea with just implementing the compose is that its likely that other services like Postgres are dependencies to the project which should just be ready to go instead of needing to install sep.